most features working. test eveything once. yet to add edit resources and event and add the members and image details to firebase

+++ lib/viewmodels/view_event_view_model.dart
@@ -1,3 +1,5 @@
+import 'package:dsc_sastra_admin/services/dialog_service.dart';
+
 import '../locator.dart';
 import '../models/event_model.dart';
 import '../services/firestore_service.dart';
@@ -5,11 +7,22 @@ import 'base_model.dart';
 
 class ViewEventViewModel extends BaseModel {
   FirestoreService _firestoreService = locator<FirestoreService>();
+  DialogService _dialogService = locator<DialogService>();
   getEvent() {
     return _firestoreService.getEvent();
   }
 
-  deleteEvent(Event event) async{
-     await _firestoreService.removeEvent(event.docid);
+  deleteEvent(Event event) async {
+    var response = await _dialogService.showConfirmationDialog(
+      title: 'Delete',
+      description: 'Are you sure?',
+      cancelTitle: 'Cancel',
+      confirmationTitle: 'Yes',
+    );
+    if (response.confirmed) {
+      await _firestoreService.removeEvent(event.docid);
+    }
+
+    
   }
 }
